Move benchmark comparison projection inward
