+++ examples/control_plane/run-compaction-readmodel-smoke.py
@@ -10,12 +10,17 @@ if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
 from loopx import status as status_module  # noqa: E402
+from loopx.control_plane.goals import goal_frontier as goal_frontier_read_model  # noqa: E402
 from loopx.control_plane.goals.goal_vision import compact_goal_vision_packet  # noqa: E402
-from loopx.control_plane.runtime import run_compaction as run_compaction_read_model  # noqa: E402
+from loopx.control_plane.runtime import (  # noqa: E402
+    benchmark_comparison as benchmark_comparison_read_model,
+)
+from loopx.control_plane.runtime import (  # noqa: E402
+    run_compaction as run_compaction_read_model,
+)
 from loopx.control_plane.runtime.run_ingest_health import (  # noqa: E402
     worker_bridge_ingest_health_note,
 )
-from loopx.control_plane.goals import goal_frontier as goal_frontier_read_model  # noqa: E402
 from loopx.session_runtime import SESSION_RUNTIME_READONLY_PROJECTION_SCHEMA_VERSION  # noqa: E402
 
 
@@ -122,8 +127,12 @@ def _direct_compact_run(run: dict[str, object]) -> dict[str, object]:
         compact_benchmark_run=status_module.compact_benchmark_run,
         worker_bridge_ingest_health_note=worker_bridge_ingest_health_note,
         compact_benchmark_result=status_module.compact_benchmark_result,
-        compact_benchmark_comparison=status_module.compact_benchmark_comparison,
-        benchmark_comparison_decision_note=status_module.benchmark_comparison_decision_note,
+        compact_benchmark_comparison=(
+            benchmark_comparison_read_model.compact_benchmark_comparison
+        ),
+        benchmark_comparison_decision_note=(
+            benchmark_comparison_read_model.benchmark_comparison_decision_note
+        ),
         compact_benchmark_learning_ledger=status_module.compact_benchmark_learning_ledger,
         compact_benchmark_experiment_report=status_module.compact_benchmark_experiment_report,
         benchmark_experiment_report_readiness_note=(
